refactor: split monolithic page.tsx into focused components
Break the 782-line Dashboard page into 12 components and 3 shared modules:
- LeftSidebar, TopBar, HQOverview, SkillDetail, SecretsPanel, RightPanel
- ScheduleEditor, SpecNode, LoadingScreen, ErrorScreen
- ImportModal, AuthModal
- lib/types.ts, lib/constants.ts, lib/utils.ts

page.tsx is now 134 lines — pure state management and composition.

Closes HYP-16
