+++ dashboard/components/SpecNode.tsx
@@ -1,0 +1,24 @@
+import type { SpecElement } from '../lib/types'
+
+export function SpecNode({ id, elements }: { id: string; elements: Record<string, SpecElement> }) {
+  const el = elements[id]; if (!el) return null
+  const p = (el.props || {}) as Record<string, string | number | boolean | string[][] | string[] | undefined>
+  const kids = el.children?.map(cid => <SpecNode key={cid} id={cid} elements={elements} />)
+  switch (el.type) {
+    case 'Card': return (<div className="card-hst p-[var(--space-md)]">{(p.title || p.description) && <div className="mb-3">{p.title && <h3 className="font-display text-lg text-primary-100">{String(p.title)}</h3>}{p.description && <p className="text-xs text-primary-50 mt-0.5">{String(p.description)}</p>}</div>}<div className="space-y-3">{kids}</div></div>)
+    case 'Stack': return <div className={`flex ${p.direction === 'horizontal' ? 'flex-row' : 'flex-col'} ${p.gap === 'lg' ? 'gap-[var(--space-lg)]' : p.gap === 'sm' ? 'gap-[var(--space-xs)]' : 'gap-[var(--space-sm)]'}`}>{kids}</div>
+    case 'Grid': return <div className={`grid gap-[var(--space-sm)]`} style={{ gridTemplateColumns: `repeat(${typeof p.columns === 'number' ? p.columns : 2}, 1fr)` }}>{kids}</div>
+    case 'Heading': { const cls = p.level === 'h1' ? 'font-display text-2xl' : p.level === 'h2' ? 'font-display text-lg' : 'font-display text-sm text-primary-70'; return <div className={cls}>{String(p.text || '')}</div> }
+    case 'Text': { const cls = p.variant === 'caption' ? 'text-micro text-primary-40' : p.variant === 'muted' ? 'text-xs text-primary-50' : p.variant === 'lead' ? 'text-sm text-primary-70' : 'text-xs text-primary-70'; return <p className={cls}>{String(p.text || '')}</p> }
+    case 'Badge': { const v = p.variant || 'default'; const cls = v === 'destructive' ? 'bg-red-50 text-eva-red border-red-200' : v === 'secondary' ? 'bg-eva-gray text-primary-50 border-[rgba(10,10,10,0.08)]' : 'bg-green-50 text-eva-green border-green-200'; return <span className={`inline-block text-[11px] px-2 py-0.5 border font-mono`}>{String(p.text || '')}</span> }
+    case 'Table': { const columns = (p.columns || []) as string[]; const rows = (p.rows || []) as string[][]; return (<div className="overflow-x-auto"><table className="w-full text-[11px] font-mono"><thead><tr>{columns.map((c, i) => <th key={i} className="text-left text-primary-40 font-medium px-2 py-1.5 border-b border-[rgba(10,10,10,0.08)]">{c}</th>)}</tr></thead><tbody>{rows.map((row, ri) => <tr key={ri}>{row.map((cell, ci) => <td key={ci} className="px-2 py-1.5 text-primary-70 border-b border-[rgba(10,10,10,0.04)]">{cell}</td>)}</tr>)}</tbody></table></div>) }
+    case 'Stat': { const trend = p.trend as string | undefined; const dc = trend === 'up' ? 'text-eva-green' : trend === 'down' ? 'text-eva-orange' : 'text-primary-50'; return (<div className="bg-eva-gray p-3">{p.label && <div className="text-label mb-1">{String(p.label)}</div>}<div className="font-display text-xl text-primary-100">{String(p.value || '')}</div>{p.delta && <div className={`text-xs font-mono ${dc}`}>{String(p.delta)}</div>}</div>) }
+    case 'Progress': { const pct = Math.min(100, (Number(p.value || 0) / Number(p.max || 100)) * 100); return (<div>{p.label && <div className="text-label mb-1.5">{String(p.label)}</div>}<div className="progress-bar"><div className="progress-bar-fill" style={{ width: `${pct}%` }} /></div></div>) }
+    case 'TweetCard': return (<div className="card-hst p-3">{p.author && <div className="flex items-center gap-1.5 mb-1"><span className="text-xs font-medium text-primary-100">{String(p.author)}</span>{p.handle && <span className="text-[11px] text-primary-40">{String(p.handle)}</span>}</div>}<p className="text-xs text-primary-70 leading-relaxed">{String(p.text || '')}</p>{(p.likes || p.retweets) && <div className="flex gap-3 mt-1.5 text-[11px] text-primary-40 font-mono">{p.likes && <span>{String(p.likes)} likes</span>}{p.retweets && <span>{String(p.retweets)} RTs</span>}</div>}</div>)
+    case 'StoryLink': return (<a href={String(p.href || '#')} target="_blank" rel="noopener noreferrer" className="block card-hst card-hst-orange p-3"><div className="text-xs text-primary-100">{String(p.title || '')}</div><div className="flex gap-2 mt-0.5 text-[11px] text-primary-40 font-mono">{p.source && <span>{String(p.source)}</span>}{p.score && <span>{String(p.score)}</span>}</div></a>)
+    case 'Link': return <a href={String(p.href || '#')} target="_blank" rel="noopener noreferrer" className="text-xs text-eva-orange hover:underline underline-offset-2 font-mono">{String(p.label || p.href || '')}</a>
+    case 'Alert': { const t = p.type || 'info'; const cls = t === 'error' ? 'border-red-200 bg-red-50 text-eva-red' : t === 'warning' ? 'border-amber-200 bg-amber-50 text-amber-700' : t === 'success' ? 'border-green-200 bg-green-50 text-eva-green' : 'border-blue-200 bg-blue-50 text-blue-700'; return (<div className={`p-3 border-2 ${cls}`}>{p.title && <div className="text-xs font-bold mb-0.5">{String(p.title)}</div>}{p.message && <div className="text-[11px] opacity-80">{String(p.message)}</div>}</div>) }
+    case 'Separator': return <div className="warning-stripes" />
+    default: return null
+  }
+}
